Refactored CalculatePartA to own module
