+++ aoc2021-day22.cabal
@@ -25,6 +25,7 @@ source-repository head
 
 library
   exposed-modules:
+      CalculatePartA
       Cuboid
       Day22
       Day22Old
@@ -51,6 +52,7 @@ library
 executable aoc2021-day22-exe
   main-is: Main.hs
   other-modules:
+      CalculatePartA
       Cuboid
       Day22
       Day22Old
@@ -85,6 +87,7 @@ test-suite aoc2021-day22-test
       SegmentSpec
       SegmentTest
       SegmentTestSpec
+      CalculatePartA
       Cuboid
       Day22
       Day22Old
